requirements.txt
flask==3.1.1
pyopenssl==25.1.0
werkzeug==3.1.3
cryptography==45.0.3
python-dotenv==1.0.1
flask-cors==4.0.0
gunicorn==23.0.0
couchdb==1.2
itsdangerous>=2.2.0
requests==2.31.0
argon2-cffi==23.1.0
